refactor(mobile): auth 모델 파일을 역할별로 분리
- auth.model.ts → user.model.ts, auth-tokens.model.ts, auth.policy.ts
- SignUpFormData 스키마를 presentations/schemas로 이동
- 도메인 모델과 프레젠테이션 레이어 관심사 분리
