+++ apps/mobile/src/features/auth/models/auth-tokens.model.ts
@@ -1,0 +1,7 @@
+export interface AuthTokens {
+  userId: string;
+  accessToken: string;
+  refreshToken: string;
+  userName: string | null;
+  userProfileImage: string | null;
+}
