+++ apps/mobile/src/features/auth/presentations/components/SignUpPasswordForm.tsx
@@ -1,5 +1,5 @@
 import { PASSWORD_RULES, passwordSchema } from '@aido/validators';
-import type { SignUpFormData } from '@src/features/auth/models/auth.model';
+import type { SignUpFormData } from '@src/features/auth/presentations/schemas/sign-up-form.schema';
 import { ANIMATION } from '@src/shared/constants/animation.constants';
 import { useStepper } from '@src/shared/hooks/useStepper';
 import { KeyboardAdaptiveButton } from '@src/shared/ui/Button';
