+++ apps/mobile/src/features/auth/services/auth.mapper.ts
@@ -1,5 +1,7 @@
 import type { AuthTokens as AuthTokensDTO, CurrentUser } from '@aido/validators';
-import { AuthPolicy, type AuthTokens, type User } from '../models/auth.model';
+import { AuthPolicy } from '../models/auth.policy';
+import type { AuthTokens } from '../models/auth-tokens.model';
+import type { User } from '../models/user.model';
 
 export const toUser = (dto: CurrentUser): User => ({
   id: dto.userId,
